add (solution): add quadratic task

+++ 01_week/tasks/quadratic/quadratic.cpp
@@ -12,34 +12,34 @@ void SolveQuadratic(int a, int b, int c) {
     else if (a == 0 && b != 0 && c == 0)
         std::cout << "0";
     else if (a == 0 && b != 0 && c != 0)
-        std::cout << std::setprecision(6) << (long double)-c / (long double)b;
+        std::cout << std::setprecision(6) << static_cast<long double>(-c) / static_cast<long double>(b);
     else if (a != 0 && b == 0 && c == 0)
         std::cout << "0";
     else if (a != 0 && b == 0 && c != 0) {
         if (c > 0)
             std::cout << "no solutions";
         else {
-            long double x = sqrt((long double)-c / (long double)a);
+            long double x = sqrt(static_cast<long double>(-c) / static_cast<long double>(a));
             std::cout << std::setprecision(6) << -x << ' ' << x;
         }   
     }
     else if (a != 0 && b != 0 && c == 0) {
-        long double x = (long double)-b / (long double)a;
+        long double x = static_cast<long double>(-b) / static_cast<long double>(a);
         if (x < 0)
             std::cout << std::setprecision(6) << x << " 0";
         else
             std::cout << std::setprecision(6) << "0 " << x;
     }
     else {
-        long long d = b*b - 4*a*c;
+        long long d = (long long)b*b - 4LL*a*c;
         if (d < 0)
             std::cout << "no solutions";
         else if (d == 0)
-            std::cout << std::setprecision(6) << (long double)-b / (2.0L*a);
+            std::cout << std::setprecision(6) << static_cast<long double>(-b) / (2.0L*a);
         else {
             long double sqrtd = sqrt(d);
-            long double x1 = ((long double)-b + sqrtd) / (2.0L*a);
-            long double x2 = ((long double)-b - sqrtd) / (2.0L*a);
+            long double x1 = (static_cast<long double>(-b) + sqrtd) / (2.0L*a);
+            long double x2 = (static_cast<long double>(-b) - sqrtd) / (2.0L*a);
             if (x1 < x2)
                 std::cout << std::setprecision(6) << x1 << ' ' << x2;
             else
